+++ inputs.tf
@@ -1,5 +1,5 @@
 variable "kube_namespace" {
-  default = ""
+  description = "The namespace where the kafka cluster will be deployed."
 }
 
 variable "kafka_replicas" {
@@ -22,10 +22,53 @@ variable "zookeeper_container_image_version" {
   default     = "5.3.0"
 }
 
-variable "bam_resource_limits" {
-  type = "map"
+variable "kafka_resource_requests" {
+  type        = "map"
+  description = <<EOF
+The requested resources for kafka from the kubernetes master. Should be a map
+with keys "cpu" and "memory".
+EOF
+
+  default = {
+    cpu    = "500m"
+    memory = "1280Mi"
+  }
 }
 
-variable "bam_resource_requests" {
-  type = "map"
+variable "kafka_resource_limits" {
+  type        = "map"
+  description = <<EOF
+The resource limit for Kafka. Brokers exceeding these limits will be evicted.
+Should be a map with keys "cpu" and "memory".
+EOF
+
+  default = {
+    cpu     = "500m"
+    memory  = "1280Mi"
+  }
+}
+
+variable "zookeeper_resource_requests" {
+  type        = "map"
+  description = <<EOF
+The requested resources for zookeeper from the kubernetes master. Should be a
+map with keys "cpu" and "memory".
+EOF
+
+  default = {
+    cpu    = "250m"
+    memory = "512Mi"
+  }
+}
+
+variable "zookeeper_resource_limits" {
+  description = <<EOF
+The resource limit for Zookeeper. Replicas exceeding these limits will be
+evicted. Should be a map with keys "cpu" and "memory".
+EOF
+
+  default = {
+    cpu    = "250m"
+    memory = "512Mi"
+  }
 }